fix: article collect in mobile mode
