+++ src/components/Article/ArticleActions.tsx
@@ -7,6 +7,7 @@ import { useEnsName } from 'wagmi';
 
 import CollectIcon from '@/assets/collect.svg';
 import { Bookmark } from '@/components/Actions/Bookmark.js';
+import { ArticleCollect } from '@/components/Article/ArticleCollect.js';
 import { ArticleShare } from '@/components/Article/ArticleShare.js';
 import { ClickableArea } from '@/components/ClickableArea.js';
 import { Tips } from '@/components/Tips/index.js';
@@ -14,8 +15,9 @@ import { Tooltip } from '@/components/Tooltip.js';
 import { Source } from '@/constants/enum.js';
 import { classNames } from '@/helpers/classNames.js';
 import { useFireflyIdentity } from '@/hooks/useFireflyIdentity.js';
+import { useIsMedium } from '@/hooks/useMediaQuery.js';
 import { useToggleArticleBookmark } from '@/hooks/useToggleArticleBookmark.js';
-import { CollectArticleModalRef } from '@/modals/controls.js';
+import { CollectArticleModalRef, DraggablePopoverRef } from '@/modals/controls.js';
 import { type Article, ArticlePlatform } from '@/providers/types/Article.js';
 
 interface ArticleActionsProps {
@@ -26,7 +28,7 @@ export const ArticleActions = memo<ArticleActionsProps>(function ArticleActions(
     const mutation = useToggleArticleBookmark();
     const identity = useFireflyIdentity(Source.Wallet, article.author.id);
     const { data: ens } = useEnsName({ address: article.author.id });
-
+    const isMedium = useIsMedium();
     return (
         <div className="flex items-center justify-between">
             <div className="text-xs leading-[24px] text-second">
@@ -42,9 +44,15 @@ export const ArticleActions = memo<ArticleActionsProps>(function ArticleActions(
                         <Tooltip content={t`Collect`} placement="top">
                             <motion.button
                                 onClick={() => {
-                                    CollectArticleModalRef.open({
-                                        article,
-                                    });
+                                    if (isMedium) {
+                                        CollectArticleModalRef.open({
+                                            article,
+                                        });
+                                    } else {
+                                        DraggablePopoverRef.open({
+                                            content: <ArticleCollect article={article} />,
+                                        });
+                                    }
                                 }}
                                 className="inline-flex h-7 w-7 items-center justify-center rounded-full hover:bg-secondarySuccess/[.20]"
                                 whileTap={{ scale: 0.9 }}
